feat: Set up the app routes

+++ db/migrate/20221010183438_create_recipes.rb
@@ -2,8 +2,8 @@ class CreateRecipes < ActiveRecord::Migration[7.0]
   def change
     create_table :recipes do |t|
       t.string :name, null: false
-      t.integer :preparation_time, null: false, default: 0
-      t.integer :cooking_time, null: false, default: 0
+      t.float :preparation_time, null: false, default: 0
+      t.float :cooking_time, null: false, default: 0
       t.text :description, null: false
       t.boolean :public, null: false, default: true
